refactor(workstation): render project tabs through UnifiedTabContent
Convert the three remaining keep-alive-trio renderers off
HostCoupledPlaceholder (project-workitems -> WorkItemsPage,
project-linear-projects / project-linear-work-items -> LinearProjectsPage),
sourcing action callbacks from the hoisted Project host context and data
from tab.data.

Rewrite ProjectManagerContentRouter to mount project tab content through
UnifiedTabContent: the keep-alive trio still renders one pane per open tab
with display:none toggling, and the active non-trio project tab routes
through the dispatcher. chat-session and git-commit-detail keep their
bespoke inline branches (the host needs ChatView secondary / direct
git-commit mount, not the unified renderers).
