+++ src/modules/WorkStation/TabContent/renderers/projectWorkitemsCompat.tsx
@@ -1,32 +1,103 @@
 /**
  * Renderer wrapper for `project-workitems` (no-hyphen variant) tabs.
  *
- * Same renderer surface as `project-work-items` today (the router
- * deduplicates via the `persistentWorkItemTabs` list). Phase 1b
- * keeps both registry entries so the union check passes; Phase 2
- * is expected to merge the spelling.
+ * Renders the project work-items surface (`WorkItemsPage`) through the unified
+ * dispatcher, pulling action callbacks from the hoisted Project host context
+ * and deriving breadcrumb / project view from tab data — mirroring how
+ * `ProjectManagerContentRouter` mounts `WorkItemsPage` for `project-workitems`
+ * tabs (part of the persistent keep-alive trio).
  *
- * Filename has a `Compat` suffix because macOS / Windows treat
- * `projectWorkitems.tsx` and `projectWorkItems.tsx` as the same
- * file (case-insensitive FS). Once the spellings are merged in
- * Phase 2 this file can be deleted.
- *
- * TODO(phase-2): merge with `project-work-items` once the persistent
- * tab cache is moved into the wrapper, then delete this file.
+ * TODO(phase-2): merge with `project-work-items` once the persistent tab cache
+ * is fully owned by the dispatcher, then delete this file.
  */
 import React, { memo } from "react";
+import { useTranslation } from "react-i18next";
+
+import { getProjectManagerBreadcrumbSegments } from "@src/modules/ProjectManager/ProjectManagerLayout/components/projectManagerRouterUtils";
+import { useProjectHostContext } from "@src/modules/ProjectManager/ProjectManagerLayout/context/projectHostContext";
+import WorkItemsPage from "@src/modules/ProjectManager/WorkItems";
+import {
+  getProjectWorkItemsTabChrome,
+  getWorkItemDetailTabChrome,
+  normalizeProjectDetailSurfaceView,
+} from "@src/store/workstation/tabs";
 
 import type { UnifiedTabContentProps } from "../types";
-import { HostCoupledPlaceholder } from "./HostCoupledPlaceholder";
 
 const ProjectWorkitemsCompatTabRenderer: React.FC<UnifiedTabContentProps> =
-  memo(({ tab }) => (
-    <HostCoupledPlaceholder
-      tabType={tab.type}
-      title={String(tab.title ?? "Work Items")}
-      hostNote="Project workitems (no-hyphen alias) still rendered by ProjectManagerContentRouter (keep-alive cache). Phase 2 will lift this through the dispatcher context and merge with project-work-items."
-    />
-  ));
+  memo(({ tab, isActive }) => {
+    const { t } = useTranslation("projects");
+    const {
+      repoPath,
+      onOpenProjects,
+      onCreateProject,
+      onCreateWorkItem,
+      onExpandWorkItemToTab,
+      onOpenChatSession,
+      onOpenRepoSettings,
+      onCloseTab,
+      onUpdateTabData,
+      onUpdateTabMeta,
+      onSetTabUnsaved,
+      onEmbeddedWorkItemDetailStateChange,
+      onProjectListRefreshRequested,
+    } = useProjectHostContext();
+
+    const projectView = normalizeProjectDetailSurfaceView(tab.data.projectView);
+
+    return (
+      <WorkItemsPage
+        projectId={tab.data.projectId as string}
+        projectName={tab.data.projectName as string}
+        cachedProjectSlug={tab.data.projectSlug as string | undefined}
+        repoPath={repoPath}
+        projectView={projectView}
+        onProjectViewChange={(nextProjectView) => {
+          onUpdateTabData(tab.id, { projectView: nextProjectView });
+        }}
+        breadcrumbSegments={getProjectManagerBreadcrumbSegments(tab, t)}
+        workStationTabId={tab.id}
+        isActive={isActive}
+        onProjectSlugResolved={(slug) => {
+          onUpdateTabData(tab.id, { projectSlug: slug });
+        }}
+        onProjectNameUpdated={(projectName) => {
+          onUpdateTabData(tab.id, { projectName });
+          onUpdateTabMeta(tab.id, getProjectWorkItemsTabChrome(projectName));
+        }}
+        onOpenProjects={onOpenProjects}
+        onCreateProject={onCreateProject}
+        onEmbeddedWorkItemDetailStateChange={(tabId, state) =>
+          onEmbeddedWorkItemDetailStateChange(
+            tabId,
+            state,
+            tab.data.projectName as string
+          )
+        }
+        onEmbeddedWorkItemNameUpdated={(workItemName) => {
+          onUpdateTabMeta(tab.id, getWorkItemDetailTabChrome(workItemName));
+        }}
+        onCreateWorkItem={onCreateWorkItem}
+        onProjectDeleted={() => {
+          onCloseTab(tab.id);
+          onProjectListRefreshRequested();
+        }}
+        onSetUnsaved={(unsaved) => onSetTabUnsaved(tab.id, unsaved)}
+        onOpenRepoSettings={() => onOpenRepoSettings("members")}
+        onOpenChatSession={onOpenChatSession}
+        onExpandWorkItemToTab={(workItemId, workItemName, pendingUpdates) =>
+          onExpandWorkItemToTab(
+            tab.data.projectId as string,
+            tab.data.projectName as string,
+            tab.data.projectSlug as string | undefined,
+            workItemId,
+            workItemName,
+            pendingUpdates
+          )
+        }
+      />
+    );
+  });
 
 ProjectWorkitemsCompatTabRenderer.displayName =
   "ProjectWorkitemsCompatTabRenderer";
